+++ ezy-e-invoice.php
@@ -17,7 +17,7 @@
 
 if ( ! defined( 'ABSPATH' ) ) exit;
 
-define( 'EZYEIN_VERSION', '2.1.9' );
+define( 'EZYEIN_VERSION', '2.2.0' );
 define( 'EZYEIN_DIR',     plugin_dir_path( __FILE__ ) );
 define( 'EZYEIN_URL',     plugin_dir_url( __FILE__ ) );